requirements.txt
pyautogui
keyboard
psutil
pywin32
pillow
pystray